# Optimize BitOperations uses in CoreLib