--- a/src/libraries/System.Private.CoreLib/src/System/Buffers/Text/FormattingHelpers.CountDigits.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Buffers/Text/FormattingHelpers.CountDigits.cs
@@ -104,7 +104,8 @@ namespace System.Buffers.Text
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static int CountHexDigits(ulong value)
         {
-            return (64 - BitOperations.LeadingZeroCount(value | 1) + 3) >> 2;
+            // The number of hex digits is Log16(value) + 1, or Log2(value) / 4 + 1
+            return (BitOperations.Log2(value) >> 2) + 1;
         }
 
         // Counts the number of trailing '0' digits in a decimal number.