--- a/src/libraries/System.Private.CoreLib/src/System/Buffers/Utilities.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Buffers/Utilities.cs
@@ -14,7 +14,13 @@ namespace System.Buffers
         internal static int SelectBucketIndex(int bufferSize)
         {
             Debug.Assert(bufferSize >= 0);
-            return BitOperations.Log2Unsafe((uint)bufferSize - 1 | 0xF) - 3;
+
+            // Buffers are bucketed so that a request between 2^(n-1) + 1 and 2^n is given a buffer of 2^n
+            // Bucket index is log2(bufferSize - 1) with the exception that buffers between 1 and 16 bytes
+            // are combined, and the index is slid down by 3 to compensate.
+            // Zero is a valid bufferSize, and it is assigned the highest bucket index so that zero-length
+            // buffers are not retained by the pool. The pool will return the Array.Empty singleton for these.
+            return BitOperations.Log2((uint)bufferSize - 1 | 15) - 3;
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]